# artifacts/api-server/openapi.yaml
openapi: 3.0.3
info:
  title: ISM Finance API
  version: 1.2.0
  description: API REST do SaaS financeiro com painel JWT e integração Power BI (v1 token).
servers:
  - url: http://localhost:3000/api
    description: Local
...
paths:
...
  /lancamentos:
    get:
      tags: [Lancamentos]
      operationId: getLancamentos
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: tipo
          schema: { type: string, enum: [CP, CR] }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: conta_id
          schema: { type: integer }
        - in: query
          name: parceiro_id
          schema: { type: integer }
        - in: query
          name: data_inicio
          schema: { type: string, format: date }
        - in: query
          name: data_fim
          schema: { type: string, format: date }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: Lista de lançamentos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopePaginated" }
    post:
      tags: [Lancamentos]
      operationId: postLancamentos
      security: [{ bearerAuth: [] }]
      description: Aplica preenchimento automático de departamento e centro de custo via parceiro quando esses campos não forem enviados.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LancamentoInput" }
      responses:
        "201":
          description: Lançamento criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
  /lancamentos/{id}:
    get:
      tags: [Lancamentos]
      operationId: getLancamentoById
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "200":
          description: Lançamento
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
    put:
      tags: [Lancamentos]
      operationId: putLancamentoById
      security: [{ bearerAuth: [] }]
      description: Também aplica preenchimento automático por parceiro quando necessário.
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LancamentoInput" }
      responses:
        "200":
          description: Lançamento atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
    delete:
      tags: [Lancamentos]
      operationId: deleteLancamentoById
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      responses:
        "200":
          description: Lançamento removido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelopeGeneric" }
        "403":
          $ref: "#/components/responses/ForbiddenError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerV1Token:
      type: http
      scheme: bearer
      description: Token estático validado na tabela `tokens_api`.
  responses:
...
    ForbiddenError:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelopeError"
  schemas:
    ApiErrorItem:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
    ApiEnvelopeGeneric:
      type: object
      properties:
        data: {}
        meta:
          nullable: true
          type: object
          additionalProperties: true
        errors:
          nullable: true
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorItem"
    ApiEnvelopePaginated:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelopeGeneric"
        - type: object
          properties:
            meta:
              type: object
              properties:
                total: { type: integer }
                page: { type: integer }
                limit: { type: integer }
...
    ApiEnvelopeError:
      type: object
      properties:
        data:
          nullable: true
        meta:
          nullable: true
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorItem"
...
    LancamentoInput:
      type: object
      required: [tipo, vencimento, valor]
      properties:
        tipo: { type: string, enum: [CP, CR] }
        vencimento: { type: string, format: date }
        competencia: { type: string, format: date }
        conta_id: { type: integer }
        parceiro_id: { type: integer }
        descricao: { type: string }
        valor: { type: number }
        status: { type: string }
        plano_conta_id: { type: integer }
        departamento_id: { type: integer, nullable: true }
        centro_custo_id: { type: integer, nullable: true }
        parcela_atual: { type: integer }
        total_parcelas: { type: integer }
        riscos:
          type: array
          items: { type: string }